+++ src/components/views/create/ProjectTips/ProjectTip.tsx
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 import { H6, IconBulbOutline24, neutralColors } from '@giveth/ui-design-system';
-import { ECreateProjectSections } from '../CreateProject';
 import { Flex } from '@/components/styled-components/Flex';
 import DefaultTip from './DefaultTip';
 import TitleTip from './TitleTip';
@@ -9,66 +8,58 @@ import BannerImageTip from './BannerImageTip';
 import CategoryTip from './CategoryTip';
 import MapTip from './MapTip';
 import AddressesTip from './AddressesTip';
+import { ECreateProjectSections } from '../types';
 
 interface IProjectTipProps {
 	activeSection: ECreateProjectSections;
 }
 
+const contentMap = {
+	[ECreateProjectSections.default]: {
+		title: 'Tips to Make a Great Project',
+		component: <DefaultTip />,
+	},
+
+	[ECreateProjectSections.name]: {
+		title: 'A Captivating Title',
+		component: <TitleTip />,
+	},
+
+	[ECreateProjectSections.description]: {
+		title: 'Describing your Project',
+		component: <DescriptionTip />,
+	},
+
+	[ECreateProjectSections.categories]: {
+		title: 'Choose the Right Category',
+		component: <CategoryTip />,
+	},
+
+	[ECreateProjectSections.location]: {
+		title: 'Put your Project on the Map',
+		component: <MapTip />,
+	},
+
+	[ECreateProjectSections.image]: {
+		title: 'Adding a Banner Image',
+		component: <BannerImageTip />,
+	},
+
+	[ECreateProjectSections.addresses]: {
+		title: 'Receiving Funding',
+		component: <AddressesTip />,
+	},
+};
+
 const ProjectTip = ({ activeSection }: IProjectTipProps) => {
-	const renderContent = () => {
-		switch (activeSection) {
-			case ECreateProjectSections.default: {
-				return {
-					title: 'Tips to Make a Great Project',
-					component: <DefaultTip />,
-				};
-			}
-			case ECreateProjectSections.name: {
-				return {
-					title: 'A Captivating Title',
-					component: <TitleTip />,
-				};
-			}
-			case ECreateProjectSections.description: {
-				return {
-					title: 'Describing your Project',
-					component: <DescriptionTip />,
-				};
-			}
-			case ECreateProjectSections.categories: {
-				return {
-					title: 'Choose the Right Category',
-					component: <CategoryTip />,
-				};
-			}
-			case ECreateProjectSections.location: {
-				return {
-					title: 'Put your Project on the Map',
-					component: <MapTip />,
-				};
-			}
-			case ECreateProjectSections.image: {
-				return {
-					title: 'Adding a Banner Image',
-					component: <BannerImageTip />,
-				};
-			}
-			case ECreateProjectSections.addresses: {
-				return {
-					title: 'Receiving Funding',
-					component: <AddressesTip />,
-				};
-			}
-		}
-	};
 	return (
 		<Container>
 			<Flex gap='16px'>
 				<IconBulbOutline24 />
-				<H6>{renderContent()?.title}</H6>
+				<H6>{contentMap[activeSection].title}</H6>
 			</Flex>
 			<br />
-			{renderContent()?.component}
+			{contentMap[activeSection].component}
 		</Container>
 	);
 };
